WaymoEnv -> ScenarioEnv
